add webpack, jest and use ES6 modules to organise code
