+++ lite/.babelrc
@@ -1,0 +1,8 @@
+{
+  "presets": [["@babel/preset-env", {"modules": false}]],
+  "env": {
+    "test": {
+      "plugins": ["transform-es2015-modules-commonjs"]
+    }
+  }
+}
