+++ lite/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "forest-lite",
+  "version": "0.3.0",
+  "description": "Minimal implementation of FOREST",
+  "private": true,
+  "scripts": {
+    "test": "jest",
+    "build": "webpack"
+  },
+  "author": "andrewgryan",
+  "license": "BSD-3-Clause",
+  "devDependencies": {
+    "@babel/preset-env": "^7.11.0",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
+    "jest": "^26.4.0",
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12"
+  }
+}
