+++ lite/templates/index.html
@@ -7,7 +7,6 @@
         {{ resources | safe }}
         <script src="https://cdn.bokeh.org/bokeh/release/bokeh-api-{{ version }}.min.js" charset="utf-8"></script>
         <script src="https://unpkg.com/redux@4.0.5/dist/redux.js" charset="utf-8"></script>
-        <script src="./static/tiling.js" charset="utf-8"></script>
         <script src="./static/script.js" charset="utf-8"></script>
         <style type="text/css" media="all">
         </style>
