+++ lite/tiling.test.js
@@ -1,0 +1,8 @@
+const tiling = require('./src/tiling.js')
+
+
+test('interp1d interpolates midpoint of line segment', () => {
+    let actual = tiling.interp1d(0, 1, 0, 1)(0.5)
+    let expected = 0.5
+    expect(actual).toEqual(expected)
+})
