+++ lite/webpack.config.js
@@ -1,0 +1,12 @@
+const path = require('path')
+
+
+module.exports = {
+    entry: {
+        script: './src/script.js'
+    },
+    output: {
+        path: path.resolve(__dirname, "static")
+    }
+}
+
